+++ server/jest.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+  testEnvironment: 'node',
+  roots: ['<rootDir>/__tests__', '<rootDir>/src'],
+  testMatch: ['**/__tests__/**/*.js', '**/?(*.)+(spec|test).js'],
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
+  collectCoverageFrom: [
+    'src/**/*.js',
+    '!src/**/*.config.js',
+    '!src/server.js'
+  ],
+  coverageDirectory: 'coverage',
+  verbose: true
+};